distro: convert {Image,Intaller}Config to methods
This commit makes the {Image,Intaller}Config methods on
the YAML based imageType struct. This helps avoiding
loading/parsing the YAML repeatly for no good reason.

Note that for now we need to keep the toplevel helpers:
`defs.{Image,Intaller}Config()` as well until everything
is converted to the generic distro.