manifest: Add support for adding files to the build pipeline
We add the ability to name the build pipeline, which lets us have
multiple pipelines. Then we add the ability to copy into a build
pipeline files from another pipeline (normally another buid pipeline).

Additionally we add the ability to create directories in a build
pipeline. This is very useful in a pipeline that is the "source" of a
copy, as we can avoid issues when the source directory doesn't
originally exist in the container the source build pipeline is based
on.