image: If using a separate build container, copy bootc customization to it
When a build container is used it generally supposed to be "similar"
to the deployed container. I.e. at least based on the same distro,
etc. However it doesn't have to be exactly the same configuration. For
example, the target container may list a custom kargs in the bootc
install config, or custom composefs/fs-verity options in the ostree
prepare-root.conf file.

To ensure that such customization in that target image is picked up we
copy these files from the target container to the build container.