manifest: Switch from bool MountUnits to blueprint.GenerateMounts enum
This changes the "MountUnits" boolean to the enum in various internal
structures and functions:
 * manifest.OSCustomizations
 * manifest.OSTreeDeploymentCustomizations
 * manifest.RawBootcImage
 * disk.GenImageKernelOptions()
 * manifest.filesystemConfigStages()

The manifest.filesystemConfigStages() change also adds support for the
new "None" stage (just returning no stages).

There are also added test cases for the above.

Note: This is generally an internal reshuffling and doesn't actually
change any behaviour yet, as the distro yaml format is still the same
(still a mount_units boolean), and there is no way yet to pass in a
GenerateMounts option to BootcImage.