distro: rename `ImgTypeLabel` to `ISOLabel`
This commit renames the variable `ImgTypeLabel` in the
`iso_label_tmpl` YAML to `ISOLabel` which reflects what
it actually is.

We should probably rethink/rework this a bit in any case
because even `{{.ISOLabel}}` is a bit strange compared to
doing somehting like `{{iso_label}}` which maps directly
to the YAML and hence is easier for a user to figure out.

Thanks to Achilleas for spotting this.