distro/rhel-10: wsl-distribution config
Write (manually) the WSL configuration file that includes things to make
using the produced image more user friendly. This includes a reference
to an icon (in `redhat-logos`) and a friendly name.

In the future this will be replaced by the `wsl-setup` package.

There's a lot of duplication here between distributions because each one
of them has a separate value for one of the keys. This might indicate
that we should use a stage but I opted not to do so as this is only
temporary.

Signed-off-by: Simon de Vlieger <supakeen@redhat.com>