image: grow the anaconda rootfs image to 10 GiB
With newer versions of RHEL 8.10 packages, the anaconda-tree for the
edge-installer has grown to be larger than 4 GiB.  This makes it
impossible to build edge-installers with osbuild-composer on RHEL 8.10.

Grow the anaconda rootfs image to 10 GiB, which should be big enough to
fit even future growth.  This shouldn't affect the size of the ISO since
it's compressed using squashfs.