distro: allow AND and easier merging in conditions
This commit tweaks the condition handling in the YAMLs so that
we can make merging easier on the YAML level.

This means that each condition now needs a unique description
but that is probably good anyway to describe the reason for
the condition.

The rhel-9 azure image type needs a logical AND for its
image config. It checks for example that the arch is x86_64
and the rhel version >= 9.6.

This extends `when:` to allow multiple keys and when
multiple keys are used they are put together as AND:

```yaml
image_types:
  "workstation-live-installer":
    ...
    image_config:
      locale: "en_US.UTF-8"
      iso_rootfs_type: "squashfs"
      conditions:
        "f40 and below uses ext4 based iso rootfs":
          when:
            version_less_than: "41"
            distro_name: "fedora"
          merge:
            iso_rootfs_type: "squashfs-ext4"
```
With that we can express all our intend nicely and we can
also express what a "when" is doing, i.e. for some (like
package sets) it will append, for others (image-types) it
will merge and for installer-config it will override. All
of this is now easier to read via:
```yaml
when: <expr>
override: ...
``

Alternative version to https://github.com/osbuild/images/pull/1600