distro: handle `centos-9` and similar versions better
This commit adds a special `versionStringForVerCmp()` that
s a special version string for our version
compare that will assume that any version with no minor is
automatically higher than any compare with a minor version.

The rational is that "centos-9" is always higher than any "rhel-9.X"
version for our version compare (centos is always "rolling").

TODO: this should become an explicit chose in "distro.yaml" but until
we have everything converted to generic.Distro accessing the properites
from an image type is very hard so we start here.

Note that this automatically fixes some issues in our current
image defintions:
1. we added "system-reinstall-bootc" to the cloud images but this
   change did not propagate to the centos-9 cloud image
2. We exlucde "microcode_ctl" in vhd for rhel >= 9.6 but this
   never propagated to centos-9
3. we remove "dnsmasq" from edge-commit for rhel < 9.6 but this
   never propagated to centos-9