image_config: restructure WSL
Renames the `WSLConfig` to `WSL` which is to contain both a `Config` and
a `DistributionConfig`; the latter in a follow-up to this commit.

Also drops the `Config` suffix from the top-level `WSLConfig` in the
image config as it's already clear that this is a config and it allows
re-using the name for the nested WSL configuration object.

Signed-off-by: Simon de Vlieger <supakeen@redhat.com>