Many: delete EOL RHEL-9 releases
Only even Y-stream releases are supported at this point. We should use
EUS repos for all that are not the latest Y-Stream release or in
development, but that will require additional work, so for
now, just delete the releases that are not supported at all.

Adjust the test manifests config map accordingly.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>