installer: Add platform packages to anaconda-tree
The anaconda-tree needs to have the platform specific packages, such as
firmware, installed, otherwise they won't be available when booting the
iso and some hardware may fail, or operate in a degraded state.

NOTE: This changes the fedora, centos, and rhel installer manifests to
include the missing firmware packages from FirmwarePackages in
BasePlatform.