test: add new `test/integration` to check local fs customizations
This commit adds a new test that checks that the file customizations
via the new `uri` is working. It cannot be added to the existing
`./test/configs/all-customizations.json` because we need a local
file that is the same on every system we run the customization
on and it has the same path. We could "hack" something into
`gen-manifests` to archive this but it seems cleaner to instead
have a real integration test.

The new integration test uses plain go and just builds a container
with a file customizations that uses the new `uri` and then checks
that the content is included in the expected place.