rhel9/azure: don't disable ResourceDisk.Format in WAAgentConfig
Remove the ResourceDisk.Format option in the WAAgentConfig (on by
default).  With this option enabled, the WAAgent takes care of writing
the dataloss warning file on temporary filesystems, so we can drop the
"manual" creation of the dataloss warning script and systemd service.

This commit also deletes the leftover files in pkg/distro/rhel/ that
were previously handling this feature.  These files implemented the
feature before the conversion of image configs to YAML.

RHEL-93894