gen-manifests: add "mock" mode for blueprint files from URIs
This commit adds a mock mode for file customizations that come
from URIs. This is needed so that we can add the a custom
file with a uri to `all-customizations.json` and still be
able to run `gen-manifests-diff` everywhere. Without the
mocking we would need a file with stable content and stable
location on every linux that `gen-manifests-diff` runs on.

The mock works by hasing the given URI and using that hash
as a filename under `cacheRoot`, for good measure it writes
the original uri into that file and then replaces the original
URI with the replacement one.

There is a new IMAGE_BUILDER_EXPERIMENTAL option
`gen-manifest-mock-bpfile-uris` to control for this. It
is not using the commandline options because that would
require that all older version of gen-manifest that are
run in `gen-manifests-diff` via `go run` support this
new commandline option. Using the env avoids this issue.