manifest: add shadow-utils when LockRoot is enabled
The lock_root_user option in the ImageConfig (enabled in the
iot-raw-image) creates a user stage in the deployment pipeline to run
usermod to lock the password.  This fails in Fedora now because
shadow-utils isn't part of the standard build root.  In #1585 we added
shadow-utils to the build root when users are defined in the image, but
the lock_root_user option doesn't add a user to the customizations.

Check for `LockRoot` as well as `Users` when deciding to include
`shadow-utils` in the build root for ostree deployments.

Fixes #1617.