distro/fedora: use new cloud-init service name in Fedora 43
cloud-init 24.3 renamed `cloud-init.service` to
`cloud-init-network.service`.  We need to enable this on Fedora 43
instead of using the old name.

This commit is a bit quick and dirty.  It's required to unblock failing
Fedora rawhide builds.  It would be a lot cleaner if we could have
conditions in different places (in the environments) but that's a bigger
undertaking and will be handled later.

[1] https://github.com/canonical/cloud-init/blob/009c4cfd30aba80699b0a014ea5010ba43661d6b/doc/rtd/reference/breaking_changes.rst#243