distro/rhel: move ansible-core to sap_extras_pkgset
Move ansible-core to the SAP extras package set.  This removes
ansible-core from the SAP Apps Azure image type.
The package is included in the SAP images for the rhel-system-roles-sap
package, so when we moved the system roles package out of the SAP apps
image type in #1612, we should have moved ansible-core too.