test: split manifest checksums into separate files
Make one file per manifest with the checksum as the contents.  This
should reduce (but not completely eliminate) merge / rebase conflicts
in the checksums when rebasing patches.
We should still expect some conflicts, when the same manifest is
modified in two branches, but we will avoid conflicts when different but
adjacent manifests are modified, which seems to happen quite often.