distro: add `ignore_image_types` for distros
For centos we need a way to exclude certain image types. We
have various options here, one is to put it into the image
types and the other to make it a property of the distro.

This commit makes it a propery of the distro because this
way the image-types are easier to reuse, i.e. when something
like "alma" is added we do not need to worry about changing
all our image types and external consumers can also define
it on the outside.