image/archive: pipeline for archive compression
Instead of using the (sadly, recently added) compression option on the
`tar` stage insert a pipeline in that compresses the archive file. This
mimicks the behavior in the `disk` image's.

Note that this is a 'step backwards' before the switch cases can be
deduplicated amongst all images that support compression.

This changes manifest checksums for the WSL image types as they were the
only ones using the tar stage *with* compression.

Signed-off-by: Simon de Vlieger <supakeen@redhat.com>