image/*: deduplicate compression switch
Instead of having a separate switch statement in each image have a
single function that returns the relevant pipeline for a given
compression.

Note that the WSL checksums change again, this is due to the removal of
SetFilename on the Archive pipeline; there's no need to set it there as
it only *really* matters for the last (exported) pipeline.

Signed-off-by: Simon de Vlieger <supakeen@redhat.com>