fedora/wsl: manually write /etc/default/locale
When a WSL image starts up the locale is unset (or set to the Windows
locale, I'm not yet sure). Then the OOBE executes, then the shell starts
but at this point we're operating with some form of locale that might
*not* be available on the system, this emits warnings.

WSL during its init checks the `/etc/default/locale` [1] file to set the
locale before executing anything. By writing this file we work around
the issue.

Signed-off-by: Simon de Vlieger <supakeen@redhat.com>