distro: drop hardcoded dracut modules from `iotInstallerImage`
The current generic `images.go:iotInstallerImage()` has
some hardcoded anaconda modules encoded. This will be
problematic when we switch other distros (like rhel-{8,9})
to the generic distro. Those are not using them.

So instead of hardcoding them make them part of the
`InstallConfig` and add them to the fedora distro.