distro: drop hardcoded presets in `iot-commit`
The `iotCommitImage` function in the generic images.go
code hardcodes some systemd presets.

This will become a problem when we move to rhel-8,9
to the generic distro. Here their "edge" images do
not use these presets. So this commit moves them into
the YAML image config instead of hardcoding them.