workload: implment workload as yaml (for rhel-8/eap7)
This commit implements a YAML version of workloads so that
we can use workloads in the rhel-8/azure-eap7-rhui.

We *could* move the workload to a package set, but that
would not give us exactly the same test manifest as before
because workloads are implemented so that they will trigger
a second resolve step which won't happen if we just add
the eapWorkload() as a package set.