distro: error if in YAML if {{.DistroVendor}} is used but not set
This commit ensures that we error if the `{{.DistroVendor}}`
template is used in a platform definition but the coresponding
`vendor:` attribute is not set in the `distros.yaml` yaml
definition of the distro.

We need to do it in a slightly roundabout way because:
a) not all distros set vendor (e.g. fedora just hardcodes it)
b) not all platforms need uefi

So we use the feature `missingkey=error` of text/template and
set the DistroVendor in the struct only if its actually
set in DistroYAML. This ensures that the template execution
fails if the template is used by the user has not set it.

Thanks to Brian!