loader: add support for platform conditionals
On rhel-8 we change the platform based on the version, i.e
ec2 images before rhel-8.9 are bios only. This needs a
new conditional.

Because it is an override and the order of conditionals is
(currently) undefined (because its a go map) it will error
if multiple conditions match. Thanks to Ondrej for this idea.