manifest: use shutil.Quote() when building strings
When building the strings for the subscription systemd
unit we need to be a bit more careful [0] with the quoting
because e.g. template names could contain strings like
```
Robert"); DROP TABLE
```
which would build into:
```
rhcConnect += fmt.Sprintf(" --content-template=\"%s\"", subscriptionOptions.TemplateName)
-> --content-template="Robert"); DROP TABLE"
```
which of course would mean that the command can no longer
run.

Note that this is only a (slightly theoretical) robustness issue,
the command injection is not a problem as the API already
allows adding arbitrary users to the image and this systemd
job runs on firstboot only. Still worth fixing if strange
names are used.

[0] AFAICT we don't do input validation for e.g. template_name,
c.f.
https://github.com/osbuild/osbuild-composer/blob/v144/internal/cloudapi/v2/openapi.v2.yml#L2295