test: add check_files_customizations() to base-host-check.sh
Because the previous approach of end-to-end testing the file
cusomizations in https://github.com/osbuild/images/pull/1157/commits/7784f3dc6b435fa03951263e48ea7cfca84c2ebd
was rejected this commit adds (basic) testing based on the
existing paradigm of "base-host-check.sh". One nice side-effect
is that it adds end-to-end testing for the existing `files`
cutomizations as a side-effect.

I initially used:
```
$ jq -r '.blueprint.customizations.files | map(select(has("uri"))) | .[] | .path'
```
to filter only the uri based ones but that seems unnecessary.

Note that this test only checks for the existance of the filesystem
customizatons target path not the content. For the simple case when
"data" is provided we could check but for the "uri" case we do not
know the content as the file usually comes from the host.  The
existing testing framework makes the content check difficult, so we
settle for this for now. There is an alternative approach in
https://github.com/osbuild/images/pull/1157/commits/7784f3dc6b435fa03951263e48ea7cfca84c2ebd
that may eventually be considered that is more direct and runs runs
locally but different from the existing paradigm so it needs further
discussion.