distro.manifest: add support to instantiate bootc manifests
This commit adds support to instanciate bootc based images
in the `imageType.Manifest()` function. This is needed
because a bootc disk image needs container SourceSpecs
instead of repositories to initialize.

Note that we could also change the interface of
ImageKind.InstantiateManifest() and add e.g.
container.SourceSpec there or provide a new struct type
inputs that contains both repositories and container
specs. But doing this is a bigger change with more churn.