distro: add new `transform_re` to the distro
This commit adds  a new `transform_re` expression to the distro
yaml so that a regexp can be used to translate a name into the
canonical "name-major.minor". This is needed for rhel-8 where
there is support to create a distro with `rhel-810` instead of
`rhel-8.10`.

This commit also includes the regexp for rhel8