distro: add edge-{commit,container} to skipTest for rhel8
This commit adds `edge-{commit,container}` to the skipTest
for TestRH8_DiskCustomizationsCheckOptions(). The reason
is that the original rhel8 checkOptions() did not check
if partitioning customizations were used (unlike e.g. rhel10)
so the original rhel8 check was incomplete, with the new
stricter check we need to skip these image types just like
we do on rhel-10.