distro,distroidparser: implement defs.ParseID
For the special cases in rhel8,rhel9 we need a version of
ParseID that comes from YAML (or it could have come from
starlark if this was a PR that would have been actually
accepted).

So this commit implements `defs.ParseID()` that will
apply any `transform_re` expressions that can be used
to create special cases for naming of distros.