distro: add support for `unsupported_partitioning_modes`
This commit adds support for unsupported_partitioning_modes
via YAML so that the RHEL-8 constraints for "edge-raw-image"
and "edge-simplified-installer" can be expressed. This was
not tested before and now a test is added.

There is a perculiar change needed in the partitioning, on
fedora (that is now the generic distro) all IoT images will
force partitioning mode "auto-lvm", this is also not showing
up in any of our tests, its not guarded via a check for
manifest.DISTRO_FEDORA and (indirectly) tested via the new
TestRH8_UnsupportedPartitioningModes() test.