generic: apply partition mode mutation only on fedora
This commit fixes an issue that we currently mutate the
partitioning mode for any images that uses rpm_ostree
to AutoLVMPartitioningMode.

This will be a problem when we move to rhel8 where we
have the opposite - for the `edge-raw-image` there we
disallow `lvm` and `auto-lvm`.

So this commit changes the partition table code to
only mutate the partitioning mode when fedora is used
and add tests (this almost went unnoticed as we had
no tests for this area and the manifest tests generate
a partition table but do not capture the partitioning
mode directly).

The mutation is problematic, ideally we would instead
just use `unsupported_partitioning_mode` for IoT on
fedora but that is a backwards incompatible change
as currently one can pass `BtrfsPartitioningMode`
to IoT on fedora and it will silently convert it to
`AutoLVMPartitioningMode`.