buildconfig: rename "buildconfig" to "blueprintload"
The package "buildconfig" originally comes from bootc-image-builder
and it started out with the idea that we could use a superset
of blueprints so originally the type was buildconfig. However
this got simplified since and the type in bib for BuildConfig
is just a images blueprint. So this commit renames the package,
removes the type BuildConfig in favor of just calling it
Blueprint and adds doc strings/cleanup around the exported
functions.