workflow: run bib container tests like resolver contaner tests
This commit changes the way the new `bib/container` tests are
run so that they are similar to the local resolver tests. It
also adds a `--fail-if-podman-missing` so that we can be sure
that the tests run in the container workflow.

To be able to share the "hasPodman()" helper the container_test.go
was moved to the "container_test" package name.

Note that the container tests cannot run in the regular workflow,
because that is already running inside a podman container and
nesting them does not work.