defs: add image types for bootc
This commit adds all bootc disk image types as YAML. The
`BootcDiskImage` is special in that it generates all
image types in the manifest anyway so this does not
fully reflect reality but its close enough so that the
images library can now work with bootc disk image types.

Note that for the ISO a bit more work is needed, here the
missing piece is that we need to pass use a custom
depsolver that operates on the container. This is already
almost possible via the "manifestgen.Depsolver" helper.
Its future work though.