rhel: drop azure-eap7-rhui special handling
By removing the eapWorload() from azure-eap7-rhui we can remove
the special cases from the rhel8 tests. When eapWorload() was
a custom workload no blueprint customizations where supported.

With the simplification in the previous commit they now are
supported. Technically this is a behavior change but it
will not break backward compatibility, just allow more now.

Note that the difference between the "workload" handling and
package sets is that workloads will not install weak dependencies
but package sets do. So this commit also excludes a bunch of
packages that are weak dependencies of the eapWorkload(). With
these excludes the resulting package list is identical between
before and after in my testing. Validated with:
```
$ gen-manifests -distros rhel-8.* -arches x86_64 -types azure-eap7-rhui
$ diff -u \
   ./test/data/manifests/rhel_8.<ver>-x86_64-azure_eap7_rhui-empty_rhel.json \
   ./test/data/manifests/rhel_8.<ver>-x86_64-azure_eap7_rhui-empty_rhel.json.good
```