many/installer: enable weak deps unconditionall
In [1] weak dependencies were disabled for all Anaconda installer types
except for the live ISO. This lead to a sharp drop in the amount of
packages included on the ISO including many firmware packages.

Fedora has reported that ISOs built with Image Builder no longer boot on
real hardware after this change. Let's revert this change to figure out
how we can do this more correctly.

[1]: https://github.com/osbuild/images/pull/1423/commits/186c814eebc1f84aaa6a0df3b8d92f1a052dbc6c

Signed-off-by: Simon de Vlieger <supakeen@redhat.com>