many: add support to generate bootc fake manifests
This commit adds support to generate fake manifests for
bootc. This a bit different from how this work for rpm
based distros. There the distro/arch/imgtype/blueprint
is enough to fully describe an image. However in the
bootc world the container is inspected for various
inputs as well so we need a (fast) way to provide
the inputs from the inspection. We could build fake
containers but it would still be a bit on the slow
side. So instead we define a new
`test/bootc-fake-containers.yaml` that can be used
to create a bunch of fake bootc inputs as if those
are read from the image itself.

With that we can (re)use the normal `config-map.json`
mechanism to get all the "bootc-*" distros and apply
the customizations.

Note that this adds some (f)ugly:
```
bootc.NewBootcForTesting()
bootc.BootcDistro.SetBuildContainerForTesting()
```
this is unavoidable if we want to use this approach
for testing. We could move to this testing to the
`booctest` fake containers approach from
https://github.com/osbuild/images/pull/1828

The downside is slightly slower fake image generation
and that we need to run this all as root because
we need "podman container mount" for the introspection
to work (this is a big downside as it will generate
root owned manifest checksums).