image_config: Add iso_boot_type
This adds support for setting the iso boot type in the image config
yaml, it includes new Unmarshal code and tests.

"grub2-uefi" is the default and sets up the iso to only boot on UEFI
using grub2.

"syslinux" sets up the iso to use syslinux for BIOS booting and grub2
for UEFI booting.

"grub2" sets up the iso to use grub2 for both the BIOS and UEFI booting.