gen-manifests: add `-buildconfig-allow-unknown` flag
This commit adds support for `-buildconfig-allow-unknown` that
allows to turn off stict unmarshal mode in json for the configs
in test/ - this is useful in `gen-manifests-diff` where we always
run the "old" gen-manifests against the new branch. If the new
code contains unknown (to the old code) fields it errors. With
this new addition we can run the "old" gen-manifest with
`-buildconfig-allow-unknown` and the new one without.

This allows us to diff manifests even if the build config changes,
like in PR#1656 where `distro.ImageOptions` grows a new
`Bootc` option (similar to OSTree).