rhel8: fix silly copy/paste mistakes in arch selection
This commit fixes silly copy copy/paste mistakes in the arch
selection for conditions. This will fix "when" conditions on
architecture when using the incorrect image types.

Note that this only happens because we are in a intermediate
state right now, once PR#1643 is merged and RHEL8 is a full
YAML based distro this kind of error can no longer happen.