anaconda: always enable the runtime module
Anaconda introduced a `Runtime` module in Fedora 39 [1]. This was noted
as 'must always be activated otherwise Anaconda crashes' and for users
of `activatable_modules` (which we use) to adapt to the change.

We never adapted the default states for modules we activate and we
didn't observe any Anaconda crashes. However, it seems that more
recently more parts of Anaconda's kickstart handling were migrated to
this module; amongst them the `%post` scriptlets [2] [3].

There's now multiple [4] [5] [6] [7] reports of `%post` scriptlets not
being executed on ISOs produced by programs using `images`.

As Anaconda describes this module as not being optional we will enable
it always. This changes all installer manifests.

[1]: https://anaconda-installer.readthedocs.io/en/latest/release-notes.html#new-runtime-module
[2]: https://github.com/rhinstaller/anaconda/pull/5739/files
[3]: https://github.com/rhinstaller/anaconda/releases/tag/anaconda-42.13
[4]: https://github.com/osbuild/bootc-image-builder/issues/968
[5]: https://github.com/fedora-iot/iot-distro/issues/97
[6]: https://github.com/rhinstaller/anaconda/discussions/6397
[7]: https://bugzilla.redhat.com/show_bug.cgi?id=2364442

Signed-off-by: Simon de Vlieger <supakeen@redhat.com>