distro: permissive RHC per rhel version in subscription service
On rhel 9 and 10 rhc's selinux policy is set managed in the post-install
script of the rpm. In the post-install script rhcd_t is set to
permissive. As osbuild executes all builds in a bubblewrap container,
selinux is not seen as enabled, and the semanage call has no
effect. This is why rhcd_t is set to permissive in the registration
script, where semange would be effective, as it is executed on
the (non-container) host.

On rhel 8 the rhcd_t domain is missing, and thus cannot be set to
permissive. Thus only set rhcd_t to permissive on rhel 9 and 10.

This only applies when a subscription is set in the ImageOptions, and
will determine whether or not `semanage permissive --add rhcd_t` is
added to the subscription script.