distro: permissive RHC per rhel version in subscription service
On rhel 8 the rhcd_t domain is missing, and thus cannot be set to
permissive. Only set rhcd_t to permissive on rhel 9 and 10.

This only applies when a subscription is set in the ImageOptions, and
will determine whether or not `semanage permissive --add rhcd_t` is
added to the subscription script.