disk: fix addPartitionsForBootMode for PPC64/s390x
This commit adds support for PPC64/s390x in the
addPartitionsForBootMode. Here we need to handle
things slightly differently from x86_64/aarch64:

We need to add the `PReP` partition type automatically
on ppc64le when disk customizations are used.

When the user provides an "incomplete" partition table in
the disk customizations (e.g. missing the bios boot partition
on x86_64) we automatically add it. This support was missing
for the ppc64le and the `PReP` partition (type "41").

Note that this change also means we no longer add an EFI
partition on s390x/ppc64le. Adding them was a side-effect
of the code in addPartitionsForBootMode() but according
to wikipedia [0] UEFI is not specificied for powerpc CPUs.

Closes: https://issues.redhat.com/browse/HMS-8884

[0] https://en.wikipedia.org/wiki/UEFI#Compatibility