distro: fix bug in `customizations` for kernel options
This commit fixes a bug in the kernel options handling
when getting the kernel options for the image types.

Both `osCustomizations` and `ostreeDeploymentCustomizations`
need fixing.

Instead of the defaultImageConfig we need to get the
inherited value from distro default image config and
image type image config when looking for the kernel
options.

Thanks to Tomáš for spotting this!