distro: unify transform/match into a single concept
This commit simplifies the concept of transform/match
into a single "match" directive. When moving to the
generic distros concept we can not assume that every
distro follows our schema of `<name>-<major>.<minor>`
and we also need to provide compatibility with older
ways of denoting rhel releases (like "rhel-89" which
means "rhel-8.9"). To accomendate this the "transform_re"
was added. But this causes a weird mix of "match" that
takes glob expression and transform that takes regexp.
Its also cumbersome to have two concepts that do almost
the same.

So this commit makes "match" do both the matching and
(if capture groups are used) the transformation. In
the simple case not much changes, e.g. fedora/rhel-10
that do not need backward compatibility are now written
as:
```re
match: 'fedora-[0-9][0-9]+'
match: 'rhel-10\.[0-9]{1,2}'
```
and rhel{8,9} are now written as:
```re
match: '(?P<name>rhel)-(?P<major>9)\.?(?P<minor>[0-9]{1,2})'
```

A match can either have no capture groups or name,major or
name,major,minor. And if a capture group is used the
provided name will be normalized to `<name>-<major>` or
`<name>-<major>.<minor>`.

Note that there is a certain overlap with
`distroidparser.Standardize()`. I think we want a followup
that deals with that, distroidparser was important in the
past but now with everything as a generic distro it seems
we might consider sunsetting it.

Thanks to Ondrej for suggesting this.