gen-manifests: use tmpdir to write mock bp files with URIs
The source URI for the file to be embedded doesn't matter, so we can use
a temporary directory for it.  This solves an issue where the manifest
generation would fail when the cacheRoot was not an absolute path, since
the blueprint customization requires absolute paths for content URIs.

Create a tmpdirRoot in main() and delete it at the end so we can reuse
it for any other temporary file creation we might need.