disk: Add support for plain partitions without filesystems
Add the capability to create raw (unformatted) partitions, for custom boot
setups that require partitions with specific GPT types but no filesystems.

The plain partition with fs_type none doesn't have to specify a mountpoint
in its customization, similarly to plain swap partitions that also don't
have formatted filesystems.

For these raw partitions, the label field can be used as a key to resolve
a specific GPT partition type GUID. If a label is not specified, the type
defaults to "data".

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>