distro: drop imageType.{default{Image,Installer}Config}}
This commit drops the redundant and confusing copies
of `defaultImageConfig` and `defaultInstallerConfig`
in the generic distro imageType. The information
is available inside `ImageTypeYAML` so just take it
from there.

Also the imageTypes `defaultImageConfig` needs to
inherit the config from the parent distro so makeing
the plain values available was actually a source for
errors (thanks to Thozza for spotting this).