distro: drop `workload` from imageType
We no longer support image types that set a workload directly,
see https://github.com/osbuild/images/pull/1660 and also there
is no concept (intentionally) for a workload to be YAML.

With this we can drop `workload` from the generic distro
`imageType` as this is always unset and we always create
a `workload.Custom{}` based on the blueprint config.

As a further step we might consider removing the `workload`
package entirely and just replace it with `ImageConfig`
(this ties into the idea of Achilleas to unify most configs
into inheritable ImageConfigs). But that is for later :)