distro: drop `ImageType.BasePartitionTable()`
The `ImageType.BasePartitionTable()` was important so that
the various "images.go" functions for the various go based
distros could generate a partition table. However now we
only have a single "generic" distro and it has access to
the base partition table without the need for the interface
(and there are only two callers). So drop it from the
`distro.ImageType` interface.