distro: add/use `DistroYAML.ID` consistently
We use(d) `distroNameVer` a lot when accessing YAML
data because that is what we had when transitioning
from fedora/rhel distro abstraction to the generic
distro. But now that everything is a generic distro
we can actually use `distro.ID` everywhere which
simplifies a lot of code and avoids (re)parsing
the `distroNameVer`.