distro: drop {Build,Payload}Pipelines from imagetypes
All image types currently define the build and payload
pipelines. But the build pipelines are the same (`["build"]`)
for all images and the only place that uses it is the sbom
writer so we can as well just use the static name there.

The payload pipelines are not used at all in the code AFAICT
and osbuild just needs the export to know what pipelines it
needs to run.

The extra complication is that the connection between the
pipeline names and the code is quite loose and the pipelines
are actually fully determined by the combination of the
image function, platform and compression and could (probably)
also be infered. But it AFAICT we don't need them at all so
YAGNI.